Move addToOrder event handler callback to itemShowView to add items to order
